requirements.txt
Django==2.0.3
django-haystack==2.8.0
djangorestframework==3.7.7
mysqlclient==1.3.12
pysolr==3.7.0
